Refactor: Centralize image preprocessing logic
Moves image preprocessing logic, including the CropInfoBar transform and normalization values, into the shared util.py module.

This ensures that the exact same preprocessing is applied to images during both training and inference.

Updates main.py and training.py to use the new shared components, fixing a bug where preprocessing was inconsistent and addressing duplicated code.
